+++ src/app/shell/rating-mode/RatingMode.scss
@@ -1,4 +1,4 @@
-@import '../../scss/variables.scss';
+@import '../../../scss/variables.scss';
 
 .mode-popup {
   @include below-header;
